add:삭제시 알림창 추가

+++ src/components/MyTripSettingModal.vue
@@ -47,11 +47,35 @@ const updateTrip = () => {
     })
 }
 const deleteTrip = () => {
-  axios.delete("http://localhost/delete-tripbyid/" + props.trip.trip_id)
+
+  Swal.fire({
+  title: "여행 기록을 지우시겠습니까?",
+  text: "되돌릴 수 없어요!",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "yes"
+}).then((result) => {
+  if (result.isConfirmed) {
+    axios.delete("http://localhost/delete-tripbyid/" + props.trip.trip_id)
     .then(({ data }) => {
-      alert("삭제했습니다.")
       refreshAttractions();
-  })
+      Swal.fire({
+      title: "삭제 성공",
+      text: "여행기록이 삭제되었습니다.",
+      icon: "success"
+      });
+    }).catch((error) => {
+      Swal.fire({
+      icon: "error",
+      title: "삭제 실패",
+      text: error +"에러가 발생되었습니다.",
+      });
+    })
+    
+  }
+});
 }
 const rollback = () => {
   tripName.value= props.trip.trip_name
